Documentation (correction)
Experimental use of Doxygen to help document the library. There are a few extra files which have been included which are not in the ToolboxC++Lib, such as the List and Vector templates. Also C++ code for SpriteAreaObj and SpriteObj. Previous documentation didn't have the .c directory included. Note the original RiscOS files have no DOS type file extension.