Update to !MinProg Example
Removed the task class into its own h and c++ files.
Moved the typedef enum event list into the task class itself.
Removed an unnecessary line of code in the typedef enum.
Removed the need to store the task handle as this is already carried out by the initialisation in main.